Semi-Finalized Appearance
This update adds the following features:
- Support for filtering and recoloring tools that can be used to search
the map for specific features and color by features or trends. These
features have not been fully implemented, but their overall appearance
is close to what we think it’ll be in our end product.
- Button and associated event/code skeleton controlling the legend box
on the bottom right corner of the page
- Algorithm for calculating what islands an object is attached to. Any
item added to the map that is tagged by lat/long will now also be
tagged with a respective island number